requirements/base.txt
Django==1.11.3
django-environ==0.4.3
djangorestframework==3.6.2
